// frontend/client-app/src/types/models/ProgramPatch.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ProgramPatch = {
    institution_id?: string;
    name?: string;
    program_code?: string;
    image_url?: string;
    description?: string;
    level?: string;
    duration?: Record<string, any>;
    delivery_mode?: Array<string>;
    fees?: Record<string, any>;
    eligibility?: Record<string, any>;
    curriculum_summary?: string;
    specializations?: Array<string>;
    url?: string;
    extensions?: Record<string, any>;
    confidence_score?: number;
};